refactor(share): move redactEmails next to the scan it exists to clear
CI failed the impact gate: "impact exceeded 85s budget (150s)". All 566 selected
tests passed — the job died on wall time, because touching src/lib/redact.ts
pulls a hub module into the impact set and selects far more of the suite than
this change needs.

The fix is also the better design, which is why it is a move rather than a
budget argument. redactEmails exists for exactly one reason: to satisfy
scanShareContent's email gate before a publish. Sharing that module's EMAIL_RE
is what makes the masking SUFFICIENT to clear the gate rather than merely
reducing the hit count — two copies of the pattern in two modules would drift,
and the failure mode of that drift is a refused publish at runtime. It also
keeps redactSecrets honest: an email is not a credential, and a transcript
rendered for a private gist still reads better with real addresses.

src/lib/redact.ts is back to its state on main, byte for byte.
