+++ apps/cli/src/commands/sessions-share.test.ts
@@ -3,8 +3,7 @@ import * as os from 'os';
 import * as path from 'path';
 import { fileURLToPath } from 'url';
 import { describe, expect, it } from 'vitest';
-import { publishToEndpoint, scanShareContent } from '../lib/share/publish.js';
-import { redactEmails } from '../lib/redact.js';
+import { publishToEndpoint, redactEmails, scanShareContent } from '../lib/share/publish.js';
 import { renderSessionHtmlDocument } from '../lib/session/share-html.js';
 import type { SessionMeta } from '../lib/session/types.js';
 import { renderSessionMarkdownDocument } from './sessions-render.js';
