+++ apps/cli/src/commands/sessions-share.ts
@@ -19,11 +19,11 @@ import * as path from 'path';
 import chalk from 'chalk';
 import type { Command } from 'commander';
 import { setHelpSections } from '../lib/help.js';
-import { knownSecretValuesFromEnv, redactEmails } from '../lib/redact.js';
+import { knownSecretValuesFromEnv } from '../lib/redact.js';
 import { discoverSessions } from '../lib/session/discover.js';
 import { renderSessionHtmlDocument } from '../lib/session/share-html.js';
 import type { SessionMeta } from '../lib/session/types.js';
-import { publishFile, type PublishOptions } from '../lib/share/publish.js';
+import { publishFile, redactEmails, type PublishOptions } from '../lib/share/publish.js';
 import { renderSessionMarkdownDocument, type ReasoningMode } from './sessions-render.js';
 import { selectSessions } from './sessions-export.js';
 import { parseAgentFilter } from './sessions.js';
