+++ apps/cli/src/lib/share/publish.ts
@@ -308,6 +308,28 @@ export interface SensitiveHit {
 const EMAIL_RE = /[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}/g;
 
 /**
+ * Mask every email address in `text`, so a page can be published without tripping
+ * {@link scanShareContent}.
+ *
+ * Lives here, beside the scanner, rather than in `lib/redact.ts`: it exists solely
+ * to satisfy this gate, and sharing {@link EMAIL_RE} is what makes the masking
+ * *sufficient* to clear it rather than merely reducing the hit count. Two copies of
+ * the pattern in two modules would drift, and the failure mode of that drift is a
+ * refused publish at runtime.
+ *
+ * Deliberately NOT part of `redactSecrets`: an email is not a credential, and a
+ * transcript rendered for a private gist or local review reads better with the real
+ * author addresses intact. It only becomes a leak once the text is PUBLISHED — a
+ * world-readable page carrying seven of them is the RUSH-2428 incident.
+ *
+ * The whole address goes, domain included — a personal domain identifies its owner
+ * as surely as the local part does, and keeping it would buy the reader nothing.
+ */
+export function redactEmails(text: string): string {
+  return text.replace(EMAIL_RE, '[EMAIL]');
+}
+
+/**
  * Credential-shaped strings that an agent routinely dumps into reports: GitHub
  * PATs, OpenAI/Anthropic/etc. API keys, AWS access keys, Slack tokens, and a
  * generic long `sk-…` / `Bearer …` form. Keep the set tight — false positives
